Refactor scene and renderers
