+++ manim/renderer/cairo_renderer.py
@@ -1,4 +1,3 @@
-import time
 import typing
 
 import numpy as np
@@ -8,8 +7,6 @@ from manim.utils.hashing import get_hash_from_play_call
 from .. import config, logger
 from ..camera.camera import Camera
 from ..mobject.mobject import Mobject
-from ..scene.scene_file_writer import SceneFileWriter
-from ..utils.exceptions import EndSceneEarlyException
 from ..utils.iterables import list_update
 
 
@@ -22,7 +19,6 @@ class CairoRenderer:
 
     def __init__(
         self,
-        file_writer_class=SceneFileWriter,
         camera_class=None,
         skip_animations=False,
         **kwargs,
@@ -30,8 +26,6 @@ class CairoRenderer:
         # All of the following are set to EITHER the value passed via kwargs,
         # OR the value stored in the global config dict at the time of
         # _instance construction_.
-        self.file_writer = None
-        self._file_writer_class = file_writer_class
         camera_cls = camera_class if camera_class is not None else Camera
         self.camera = camera_cls()
         self._original_skipping_status = skip_animations
@@ -41,72 +35,19 @@ class CairoRenderer:
         self.time = 0
         self.static_image = None
 
-    def init_scene(self, scene):
-        self.file_writer = self._file_writer_class(
-            self,
-            scene.__class__.__name__,
-        )
-
-    def play(self, scene, *args, **kwargs):
-        # Reset skip_animations to the original state.
-        # Needed when rendering only some animations, and skipping others.
-        self.skip_animations = self._original_skipping_status
-        self.update_skipping_status()
-
-        scene.compile_animation_data(*args, **kwargs)
-
-        if self.skip_animations:
-            logger.debug(f"Skipping animation {self.num_plays}")
-            hash_current_animation = None
-            self.time += scene.duration
-        else:
-            if config["disable_caching"]:
-                logger.info("Caching disabled.")
-                hash_current_animation = f"uncached_{self.num_plays:05}"
-            else:
-                hash_current_animation = get_hash_from_play_call(
-                    scene,
-                    self.camera,
-                    scene.animations,
-                    scene.mobjects,
-                )
-                if self.file_writer.is_already_cached(hash_current_animation):
-                    logger.info(
-                        f"Animation {self.num_plays} : Using cached data (hash : %(hash_current_animation)s)",
-                        {"hash_current_animation": hash_current_animation},
-                    )
-                    self.skip_animations = True
-                    self.time += scene.duration
-        # adding None as a partial movie file will make file_writer ignore the latter.
-        self.file_writer.add_partial_movie_file(hash_current_animation)
-        self.animations_hashes.append(hash_current_animation)
-        logger.debug(
-            "List of the first few animation hashes of the scene: %(h)s",
-            {"h": str(self.animations_hashes[:5])},
-        )
-
-        # Save a static image, to avoid rendering non moving objects.
-        self.static_image = self.save_static_frame_data(scene, scene.static_mobjects)
-
-        self.file_writer.begin_animation(not self.skip_animations)
-        scene.begin_animations()
-        if scene.is_current_animation_frozen_frame():
-            self.update_frame(scene)
-            # self.duration stands for the total run time of all the animations.
-            # In this case, as there is only a wait, it will be the length of the wait.
-            self.freeze_current_frame(scene.duration)
-        else:
-            scene.play_internal()
-        self.file_writer.end_animation(not self.skip_animations)
-
-        self.num_plays += 1
+    def init_scene(self):
+        pass
 
     def update_frame(  # TODO Description in Docstring
         self,
-        scene,
-        mobjects=None,
+        moving_mobjects,
+        skip_animations=False,
         include_submobjects=True,
         ignore_skipping=True,
+        mobjects=None,
+        meshes=None,
+        file_writer=None,
+        foreground_mobjects=None,
         **kwargs,
     ):
         """Update the frame.
@@ -126,12 +67,12 @@ class CairoRenderer:
         **kwargs
 
         """
-        if self.skip_animations and not ignore_skipping:
+        if skip_animations and not ignore_skipping:
             return
-        if not mobjects:
-            mobjects = list_update(
-                scene.mobjects,
-                scene.foreground_mobjects,
+        if not moving_mobjects:
+            moving_mobjects = list_update(
+                mobjects,
+                foreground_mobjects,
             )
         if self.static_image is not None:
             self.camera.set_frame_to_background(self.static_image)
@@ -139,11 +80,25 @@ class CairoRenderer:
             self.camera.reset()
 
         kwargs["include_submobjects"] = include_submobjects
-        self.camera.capture_mobjects(mobjects, **kwargs)
+        self.camera.capture_mobjects(moving_mobjects, **kwargs)
 
-    def render(self, scene, time, moving_mobjects, skip_animations=False):
-        self.update_frame(scene, moving_mobjects)
-        self.add_frame(self.get_frame(), skip_animations=skip_animations)
+    def render(
+        self,
+        time,
+        moving_mobjects,
+        skip_animations=False,
+        mobjects=None,
+        meshes=None,
+        file_writer=None,
+        foreground_mobjects=None,
+    ):
+        self.update_frame(
+            moving_mobjects,
+            skip_animations == skip_animations,
+            mobjects=mobjects,
+            foreground_mobjects=foreground_mobjects,
+        )
+        self.add_frame(self.get_frame(), file_writer, skip_animations=skip_animations)
 
     def get_frame(self):
         """
@@ -157,7 +112,7 @@ class CairoRenderer:
         """
         return np.array(self.camera.pixel_array)
 
-    def add_frame(self, frame, num_frames=1, skip_animations=False):
+    def add_frame(self, frame, file_writer, num_frames=1, skip_animations=False):
         """
         Adds a frame to the video_file_stream
 
@@ -173,7 +128,7 @@ class CairoRenderer:
             return
         self.time += num_frames * dt
         for _ in range(num_frames):
-            self.file_writer.write_frame(frame)
+            file_writer.write_frame(frame)
 
     def freeze_current_frame(self, duration: float):
         """Adds a static frame to the movie for a given duration. The static frame is the current frame.
@@ -199,8 +154,9 @@ class CairoRenderer:
 
     def save_static_frame_data(
         self,
-        scene,
         static_mobjects: typing.Iterable[Mobject],
+        mobjects=None,
+        foreground_mobjects=None,
     ) -> typing.Iterable[Mobject]:
         """Compute and save the static frame, that will be reused at each frame to avoid to unecesseraly computer
         static mobjects.
@@ -220,45 +176,11 @@ class CairoRenderer:
         if not static_mobjects:
             self.static_image = None
             return
-        self.update_frame(scene, mobjects=static_mobjects)
+        self.update_frame(
+            static_mobjects, mobjects=mobjects, foreground_mobjects=foreground_mobjects
+        )
         self.static_image = self.get_frame()
         return self.static_image
-
-    # def update_skipping_status(self):
-    #     """
-    #     This method is used internally to check if the current
-    #     animation needs to be skipped or not. It also checks if
-    #     the number of animations that were played correspond to
-    #     the number of animations that need to be played, and
-    #     raises an EndSceneEarlyException if they don't correspond.
-    #     """
-    #     if config["save_last_frame"]:
-    #         self.skip_animations = True
-    #     if (
-    #         config["from_animation_number"]
-    #         and self.num_plays < config["from_animation_number"]
-    #     ):
-    #         self.skip_animations = True
-    #     if (
-    #         config["upto_animation_number"]
-    #         and self.num_plays > config["upto_animation_number"]
-    #     ):
-    #         self.skip_animations = True
-    #         raise EndSceneEarlyException()
-
-    # def scene_finished(self, scene):
-    #     # If no animations in scene, render an image instead
-    #     if self.num_plays:
-    #         self.file_writer.finish()
-    #     elif config.write_to_movie:
-    #         config.save_last_frame = True
-    #         config.write_to_movie = False
-    #     else:
-    #         self.update_frame(scene)
-    #
-    #     if config["save_last_frame"]:
-    #         self.update_frame(scene)
-    #         self.file_writer.save_final_image(self.camera.get_image())
 
     def should_save_last_frame(self, num_plays):
         return config["save_last_frame"] or num_plays == 0
